requirements.txt
Flask
Flask-Babel
Flask-Cors
Flask-SQLAlchemy
python-dotenv
gunicorn
psycopg2-binary
requests    